refactor(coyote): move all business logic to jwp package
- DIP with Servlet interface
