+++ tomcat/src/main/java/nextstep/Application.java
@@ -1,6 +1,7 @@
 package nextstep;
 
 import org.apache.catalina.startup.Tomcat;
+import nextstep.jwp.servlet.CustomServlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -11,6 +12,6 @@ public class Application {
     public static void main(String[] args) {
         log.info("web server start.");
         final var tomcat = new Tomcat();
-        tomcat.start();
+        tomcat.start(new CustomServlet());
     }
 }
