+++ tomcat/src/main/java/nextstep/jwp/servlet/CustomServlet.java
@@ -1,10 +1,11 @@
-package org.apache.coyote.servlet;
+package nextstep.jwp.servlet;
 
+import org.apache.coyote.servlet.Servlet;
 import org.apache.coyote.support.HttpException;
 import org.apache.coyote.request.HttpRequest;
 import org.apache.coyote.response.HttpResponse;
 
-public class CustomServlet {
+public class CustomServlet implements Servlet {
 
     private final HttpRequestHandler requestMapper = new HttpRequestHandler();
