+++ tomcat/src/main/java/nextstep/jwp/servlet/HttpRequestHandler.java
@@ -1,20 +1,20 @@
-package org.apache.coyote.servlet;
+package nextstep.jwp.servlet;
 
 import nextstep.jwp.controller.AuthController;
+import nextstep.jwp.support.ResourceRegistry;
 import org.apache.coyote.request.HttpRequest;
 import org.apache.coyote.support.HttpException;
 import org.apache.coyote.response.HttpResponse;
-import org.apache.coyote.response.ResourceView;
 import org.apache.coyote.support.HttpMethod;
 
 public class HttpRequestHandler {
 
     private final AuthController authController = new AuthController();
-    private final ResourceView resourceView = new ResourceView();
+    private final ResourceRegistry resourceRegistry = new ResourceRegistry();
 
     public HttpResponse handle(HttpRequest request) {
         if (request.isMethodOf(HttpMethod.GET)) {
-            return resourceView.findStaticResource(request.getUri());
+            return resourceRegistry.findStaticResource(request.getUri());
         }
         if (request.getUri().matches("/register") && request.isMethodOf(HttpMethod.POST)) {
             return authController.register(request);
@@ -26,6 +26,6 @@ public class HttpRequestHandler {
     }
 
     public HttpResponse handle(HttpException exception) {
-        return resourceView.findErrorPage(exception);
+        return resourceRegistry.findErrorPage(exception);
     }
 }
