+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.connector;
 
 import org.apache.coyote.http11.Http11Processor;
-import org.apache.coyote.servlet.CustomServlet;
+import org.apache.coyote.servlet.Servlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,15 +17,15 @@ public class Connector implements Runnable {
     private static final int DEFAULT_PORT = 8080;
     private static final int DEFAULT_ACCEPT_COUNT = 100;
 
-    private final CustomServlet servlet;
+    private final Servlet servlet;
     private final ServerSocket serverSocket;
     private boolean stopped;
 
-    public Connector(CustomServlet servlet) {
+    public Connector(Servlet servlet) {
         this(servlet, DEFAULT_PORT, DEFAULT_ACCEPT_COUNT);
     }
 
-    public Connector(final CustomServlet servlet, final int port, final int acceptCount) {
+    public Connector(final Servlet servlet, final int port, final int acceptCount) {
         this.servlet = servlet;
         this.serverSocket = createServerSocket(port, acceptCount);
         this.stopped = false;
