+++ tomcat/src/main/java/org/apache/catalina/startup/Tomcat.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.startup;
 
 import org.apache.catalina.connector.Connector;
-import org.apache.coyote.servlet.CustomServlet;
+import org.apache.coyote.servlet.Servlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -11,8 +11,8 @@ public class Tomcat {
 
     private static final Logger log = LoggerFactory.getLogger(Tomcat.class);
 
-    public void start() {
-        final var connector = new Connector(new CustomServlet());
+    public void start(Servlet servlet) {
+        final var connector = new Connector(servlet);
         connector.start();
 
         try {
