+++ tomcat/src/main/java/org/apache/coyote/servlet/Servlet.java
@@ -1,0 +1,9 @@
+package org.apache.coyote.servlet;
+
+import org.apache.coyote.request.HttpRequest;
+import org.apache.coyote.response.HttpResponse;
+
+public interface Servlet {
+
+    HttpResponse service(HttpRequest request);
+}
